+++ lib/spout/commands/project_generator.rb
@@ -7,6 +7,7 @@ TEMPLATES_DIRECTORY = File.expand_path('../../templates', __FILE__)
 
 module Spout
   module Commands
+    # Generates folder and file structure for a new spout data dictionary.
     class ProjectGenerator
       def initialize(argv)
         generate_folder_structure!(argv)
@@ -31,8 +32,11 @@ EOT
         copy_file 'ruby-version', '.ruby-version'
         copy_file 'travis.yml', '.travis.yml'
         copy_file 'spout.yml', '.spout.yml'
+        copy_file 'CHANGELOG.md'
         copy_file 'Gemfile'
         copy_file 'Rakefile'
+        copy_file 'README.md'
+        copy_file 'VERSION'
         directory 'domains'
         copy_file 'keep', 'domains/.keep'
         directory 'variables'
@@ -43,9 +47,9 @@ EOT
         copy_file 'test/dictionary_test.rb'
         copy_file 'test/test_helper.rb'
         unless skip_gemfile
-          puts "         run".colorize( :green ) + "  bundle install".colorize( :light_cyan )
+          puts '         run'.colorize(:green) + '  bundle install'.colorize(:light_cyan)
           Dir.chdir(@full_path)
-          system "bundle install"
+          system 'bundle install'
         end
       end
 
@@ -55,16 +59,15 @@ EOT
         file_name = template_file if file_name == ''
         file_path = File.join(@full_path, file_name)
         template_file_path = File.join(TEMPLATES_DIRECTORY, template_file)
-        puts "      create".colorize( :green ) + "  #{file_name}"
+        puts '      create'.colorize(:green) + "  #{file_name}"
         FileUtils.copy(template_file_path, file_path)
       end
 
       def directory(directory_name)
         directory_path = File.join(@full_path, directory_name)
-        puts "      create".colorize( :green ) + "  #{directory_name}"
+        puts '      create'.colorize(:green) + "  #{directory_name}"
         FileUtils.mkpath(directory_path)
       end
-
     end
   end
 end