+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **General Changes**
   - Spout now provides a warning and skips columns in CSVs with blank headers
+  - `spout new` command now generates `CHANGELOG.md`, `README.md`, and `VERSION`
+    placeholder files
 - **Exporter Changes**
   - The export command now exports the variable forms attribute
 - **Importer Changes**